[Building] Add size and shape type init.

+++ model/BasicFuncs.h
@@ -6,9 +6,11 @@
 /**
  * @file BasicFuncs.h
  */
- 
+
 #ifndef BASIC_FUNC_H
 #define BASIC_FUNC_H
+
+#include <cstdlib>
 
 /**
  * @brief Returns random number from range [low_boundary, top_boundary].
@@ -32,7 +34,7 @@ static inline double doubleRand(int top_boundary)
  * @brief Returns random number from range [low_boundary, top_boundary]
  *        (implementation for doubles).
  */
-static inline int randFromRange(double low_boundary, double top_boundary)
+static inline double randFromRange(double low_boundary, double top_boundary)
 {
     return doubleRand(top_boundary - low_boundary) + low_boundary;
 }
